Refactored session handler'
