+++ lib/logout.js
@@ -4,7 +4,8 @@ var DOMParser             = require('xmldom').DOMParser;
 var utils                 = require('./utils');
 var trim_xml              = require('./trim_xml');
 var signers               = require('./signers');
-var InMemoryStore         = require('./in_memory_store');
+var InMemoryStore         = require('./store/in_memory_store');
+var SessionHandler        = require('./sessionHandler');
 var zlib                  = require('zlib');
 var qs                    = require('querystring');
 var util                  = require('util');
@@ -17,8 +18,7 @@ var BINDINGS = {
 
 // TODO: Lets default for HTTP-Post binding (for compatibility purpose)
 module.exports.logout = function (options) {
-  options.getSessions = options.getSessions || function(cb){ return cb(null, []); };
-  options.clearSession = options.clearSession || function(cb) { return cb(); };
+  options.sessionHandler = options.sessionHandler || new SessionHandler();
   options.store = options.store || new InMemoryStore();
 
   function prepareAndSendLogoutRequest(sessions, req, res, next) {
@@ -42,7 +42,7 @@ module.exports.logout = function (options) {
   }
 
   function finalize (req, res, next) {
-    options.clearSession(function (err) {
+    options.sessionHandler.clearIdPSession(function (err) {
       if (err) return next(err);
 
       options.store.get(function (err, data) {
@@ -153,7 +153,7 @@ module.exports.logout = function (options) {
           options.store.store(requestData, function (err) {
             if (err) { return next(err); }
 
-            options.getSessions(function (err, sessions) {
+            options.sessionHandler.getActiveSessions(function (err, sessions) {
               if (err) { return next(err); }
 
               // We remove the session from the LogoutRequest Originator.
@@ -170,7 +170,7 @@ module.exports.logout = function (options) {
       // SP Initiated flow in progress, incoming SAMLResponse from SP. Scenario 1 Step 2
       } else if (req.query.SAMLResponse || req.body.SAMLResponse) {
         // If there are sessions left, keep sending LogoutRequest to Session Participants. If not finish
-        options.getSessions(function (err, sessions) {
+        options.sessionHandler.getActiveSessions(function (err, sessions) {
           if (err) { return next(err); }
           validateSamlResponse(req, sessions, function (err, logoutResponse) {
             if (err) {
@@ -187,7 +187,7 @@ module.exports.logout = function (options) {
 
       // IDP initated - Start flow - In this case we will show a 200 when complete
       } else {
-        options.getSessions(function (err, sessions) {
+        options.sessionHandler.getActiveSessions(function (err, sessions) {
           if (err) { return next(err); }
 
           prepareAndSendLogoutRequest(sessions, req, res, next);
