+++ lib/store/in_memory_store.js
